+++ src/ui/views/wallet.rs
@@ -1,14 +1,10 @@
 //! Screens and forms related to wallet management.
 
-use dpp::{
-    dashcore::psbt::serialize::Serialize, platform_value::string_encoding::Encoding,
-    prelude::Identifier,
-};
-use itertools::Itertools;
+use dpp::dashcore::psbt::serialize::Serialize;
 
 mod add_identity_key;
 
-use std::{collections::HashSet, ops::Deref};
+use std::ops::Deref;
 
 use tuirealm::{
     event::{Key, KeyEvent, KeyModifiers},
@@ -16,16 +12,14 @@ use tuirealm::{
     Frame,
 };
 
-use self::add_identity_key::AddIdentityKeyFormController;
 use crate::{
     backend::{
-        identities::IdentityTask, state::IdentityPrivateKeysMap, AppState, AppStateUpdate,
-        BackendEvent, Task, Wallet, WalletTask,
+        identities::IdentityTask, AppState, AppStateUpdate, BackendEvent, Task, Wallet, WalletTask,
     },
     ui::{
         form::{
-            parsers::DefaultTextInputParser, ComposedInput, Field, FormController, FormStatus,
-            Input, InputStatus, SelectInput, TextInput,
+            parsers::DefaultTextInputParser, FormController, FormStatus, Input, InputStatus,
+            TextInput,
         },
         screen::{
             info_display::display_info, utils::impl_builder, widgets::info::Info, ScreenCommandKey,
@@ -35,23 +29,16 @@ use crate::{
     Event,
 };
 
-const WALLET_LOADED_COMMANDS: [ScreenCommandKey; 8] = [
+const WALLET_LOADED_COMMANDS: [ScreenCommandKey; 4] = [
     ScreenCommandKey::new("b", "Refresh wallet utxos and balance"),
     ScreenCommandKey::new("c", "Copy Receive Address"),
-    ScreenCommandKey::new("i", "Register identity"),
-    ScreenCommandKey::new("s", "Switch identity"),
     ScreenCommandKey::new("u", "Get more utxos"),
     ScreenCommandKey::new("C-w", "Clear loaded wallet"),
-    ScreenCommandKey::new("m", "Load masternode identity"),
-    ScreenCommandKey::new("l", "Load identity with private key(s)"),
 ];
 
-const IDENTITY_LOADED_COMMANDS: [ScreenCommandKey; 5] = [
+const IDENTITY_LOADED_COMMANDS: [ScreenCommandKey; 2] = [
     ScreenCommandKey::new("r", "Identity refresh"),
-    ScreenCommandKey::new("w", "Withdraw balance"),
-    ScreenCommandKey::new("d", "Copy Identity ID"),
-    ScreenCommandKey::new("k", "Add Identity key"),
-    ScreenCommandKey::new("C-e", "Clear loaded identity"),
+    ScreenCommandKey::new("w", "Withdraw identity balance to wallet"),
 ];
 
 #[memoize::memoize]
@@ -59,7 +46,6 @@ fn join_commands(
     wallet_loaded: bool,
     identity_loaded: bool,
     identity_registration_in_progress: bool,
-    identity_top_up_in_progress: bool,
 ) -> &'static [ScreenCommandKey] {
     let mut commands = vec![ScreenCommandKey::new("q", "Back to Main")];
 
@@ -67,11 +53,6 @@ fn join_commands(
         commands.extend_from_slice(&WALLET_LOADED_COMMANDS);
         if identity_loaded {
             commands.extend_from_slice(&IDENTITY_LOADED_COMMANDS);
-            if identity_top_up_in_progress {
-                commands.push(ScreenCommandKey::new("t", "Continue identity top up"));
-            } else {
-                commands.push(ScreenCommandKey::new("t", "Identity top up"));
-            }
         } else {
             if identity_registration_in_progress {
                 commands.push(ScreenCommandKey::new("i", "Continue identity registration"));
@@ -93,8 +74,6 @@ pub(crate) struct WalletScreenController {
     wallet_loaded: bool,
     identity_loaded: bool,
     identity_registration_in_progress: bool,
-    identity_top_up_in_progress: bool,
-    known_identities: Vec<Identifier>,
 }
 
 impl_builder!(WalletScreenController);
@@ -107,21 +86,14 @@ impl WalletScreenController {
             wallet_loaded,
             identity_loaded,
             identity_registration_in_progress,
-            identity_top_up_in_progress,
         ) = if let Some(wallet) = app_state.loaded_wallet.lock().await.as_ref() {
             if let Some(identity) = app_state.loaded_identity.lock().await.as_ref() {
-                let identity_top_up_in_progress = app_state
-                    .identity_asset_lock_private_key_in_top_up
-                    .lock()
-                    .await
-                    .is_some();
                 (
                     Info::new_fixed(&display_wallet(wallet)),
                     Info::new_fixed(&display_info(identity)),
                     true,
                     true,
                     false,
-                    identity_top_up_in_progress,
                 )
             } else {
                 let identity_registration_in_progress = app_state
@@ -131,29 +103,21 @@ impl WalletScreenController {
                     .is_some();
                 (
                     Info::new_fixed(&display_wallet(wallet)),
-                    Info::new_fixed(""),
+                    Info::new_fixed("No identity loaded yet. Go to Identities screen to load one."),
                     true,
                     false,
                     identity_registration_in_progress,
-                    false,
                 )
             }
         } else {
             (
                 Info::new_fixed("Wallet management commands\n\nNo wallet loaded yet"),
-                Info::new_fixed(""),
-                false,
+                Info::new_fixed("No identity loaded yet. Go to Identities screen to load one."),
                 false,
                 false,
                 false,
             )
         };
-
-        let known_identities_map = app_state.known_identities.lock().await;
-        let known_identities = known_identities_map
-            .iter()
-            .map(|(k, _)| k.clone())
-            .collect_vec();
 
         Self {
             wallet_info,
@@ -161,8 +125,6 @@ impl WalletScreenController {
             wallet_loaded,
             identity_loaded,
             identity_registration_in_progress,
-            identity_top_up_in_progress,
-            known_identities: known_identities,
         }
     }
 }
@@ -186,7 +148,6 @@ impl ScreenController for WalletScreenController {
             self.wallet_loaded,
             self.identity_loaded,
             self.identity_registration_in_progress,
-            self.identity_top_up_in_progress,
         )
     }
 
@@ -233,13 +194,6 @@ impl ScreenController for WalletScreenController {
             },
 
             Event::Key(KeyEvent {
-                code: Key::Char('s'),
-                modifiers: KeyModifiers::NONE,
-            }) if self.wallet_loaded => ScreenFeedback::Form(Box::new(
-                SwitchIdentityFormController::new(self.known_identities.clone()),
-            )),
-
-            Event::Key(KeyEvent {
                 code: Key::Char('w'),
                 modifiers: KeyModifiers::NONE,
             }) if self.identity_loaded => {
@@ -247,37 +201,9 @@ impl ScreenController for WalletScreenController {
             }
 
             Event::Key(KeyEvent {
-                code: Key::Char('i'),
-                modifiers: KeyModifiers::NONE,
-            }) if !self.identity_registration_in_progress => {
-                ScreenFeedback::Form(Box::new(RegisterIdentityFormController::new()))
-            }
-
-            Event::Key(KeyEvent {
-                code: Key::Char('i'),
-                modifiers: KeyModifiers::NONE,
-            }) if self.identity_registration_in_progress => ScreenFeedback::Task {
-                task: Task::Identity(IdentityTask::ContinueRegisteringIdentity),
-                block: true,
-            },
-
-            Event::Key(KeyEvent {
-                code: Key::Char('g'),
-                modifiers: KeyModifiers::NONE,
-            }) if self.identity_registration_in_progress => ScreenFeedback::Task {
-                task: Task::Identity(IdentityTask::ClearRegistrationOfIdentityInProgress),
-                block: true,
-            },
-
-            Event::Key(KeyEvent {
                 code: Key::Char('u'),
                 modifiers: KeyModifiers::NONE,
             }) => ScreenFeedback::Form(Box::new(SplitUTXOsFormController::new())),
-
-            Event::Key(KeyEvent {
-                code: Key::Char('t'),
-                modifiers: KeyModifiers::NONE,
-            }) => ScreenFeedback::Form(Box::new(TopUpIdentityFormController::new())),
 
             Event::Key(KeyEvent {
                 code: Key::Char('c'),
@@ -288,37 +214,6 @@ impl ScreenController for WalletScreenController {
             },
 
             Event::Key(KeyEvent {
-                code: Key::Char('d'),
-                modifiers: KeyModifiers::NONE,
-            }) if self.identity_loaded => ScreenFeedback::Task {
-                task: Task::Identity(IdentityTask::CopyIdentityId),
-                block: true,
-            },
-
-            Event::Key(KeyEvent {
-                code: Key::Char('k'),
-                modifiers: KeyModifiers::NONE,
-            }) if self.identity_loaded => {
-                ScreenFeedback::Form(Box::new(AddIdentityKeyFormController::new()))
-            }
-
-            Event::Key(KeyEvent {
-                code: Key::Char('l'),
-                modifiers: KeyModifiers::NONE,
-            }) => ScreenFeedback::Form(Box::new(LoadIdentityByIdFormController::new())),
-
-            Event::Key(KeyEvent {
-                code: Key::Char('e'),
-                modifiers: KeyModifiers::CONTROL,
-            }) if self.identity_loaded => {
-                self.identity_loaded = false;
-                ScreenFeedback::Task {
-                    task: Task::Identity(IdentityTask::ClearLoadedIdentity),
-                    block: false,
-                }
-            }
-
-            Event::Key(KeyEvent {
                 code: Key::Char('w'),
                 modifiers: KeyModifiers::CONTROL,
             }) if self.wallet_loaded => {
@@ -327,13 +222,6 @@ impl ScreenController for WalletScreenController {
                     task: Task::Wallet(WalletTask::ClearLoadedWallet),
                     block: false,
                 }
-            }
-
-            Event::Key(KeyEvent {
-                code: Key::Char('m'),
-                modifiers: KeyModifiers::NONE,
-            }) if self.wallet_loaded => {
-                ScreenFeedback::Form(Box::new(LoadMasternodeIdentityFormController::new()))
             }
 
             Event::Backend(
@@ -386,12 +274,6 @@ impl ScreenController for WalletScreenController {
                 task: Task::Identity(IdentityTask::Refresh),
                 block: true,
             },
-
-            Event::Backend(BackendEvent::TaskCompletedStateChange {
-                task: Task::Identity(IdentityTask::LoadIdentityById(_)),
-                execution_result: Ok(_),
-                app_state_update: _,
-            }) => ScreenFeedback::Form(Box::new(AddPrivateKeysFormController::new())),
 
             Event::Backend(BackendEvent::TaskCompletedStateChange {
                 execution_result,
@@ -534,50 +416,6 @@ impl FormController for SplitUTXOsFormController {
     }
 }
 
-struct SwitchIdentityFormController {
-    input: SelectInput<Identifier>,
-}
-
-impl SwitchIdentityFormController {
-    fn new(known_identities: Vec<Identifier>) -> Self {
-        Self {
-            input: SelectInput::new(known_identities),
-        }
-    }
-}
-
-impl FormController for SwitchIdentityFormController {
-    fn on_event(&mut self, event: KeyEvent) -> FormStatus {
-        match self.input.on_event(event) {
-            InputStatus::Done(id) => FormStatus::Done {
-                task: Task::Identity(IdentityTask::LoadKnownIdentity(id)),
-                block: false,
-            },
-            status => status.into(),
-        }
-    }
-
-    fn form_name(&self) -> &'static str {
-        "Switch identity"
-    }
-
-    fn step_view(&mut self, frame: &mut Frame, area: Rect) {
-        self.input.view(frame, area)
-    }
-
-    fn step_name(&self) -> &'static str {
-        "Base58 Identity ID"
-    }
-
-    fn step_index(&self) -> u8 {
-        0
-    }
-
-    fn steps_number(&self) -> u8 {
-        1
-    }
-}
-
 fn display_wallet(wallet: &Wallet) -> String {
     match wallet {
         Wallet::SingleKeyWallet(single_key_wallet) => {
@@ -590,154 +428,6 @@ fn display_wallet(wallet: &Wallet) -> String {
             let utxo_count = single_key_wallet.utxos.len();
             format!("{}\nNumber of UTXOs: {}", description, utxo_count)
         }
-    }
-}
-
-struct LoadMasternodeIdentityFormController {
-    input: ComposedInput<(
-        Field<TextInput<DefaultTextInputParser<String>>>,
-        Field<TextInput<DefaultTextInputParser<String>>>,
-    )>,
-}
-
-impl LoadMasternodeIdentityFormController {
-    fn new() -> Self {
-        Self {
-            input: ComposedInput::new((
-                Field::new(
-                    "proTxHash",
-                    TextInput::new("Enter Evonode proTxHash in base58 format"),
-                ),
-                Field::new(
-                    "Private Key",
-                    TextInput::new("Enter the Evonode private key in WIF format"),
-                ),
-            )),
-        }
-    }
-}
-
-impl FormController for LoadMasternodeIdentityFormController {
-    fn on_event(&mut self, event: KeyEvent) -> FormStatus {
-        match self.input.on_event(event) {
-            InputStatus::Done((pro_tx_hash, private_key)) => FormStatus::Done {
-                task: Task::Identity(IdentityTask::LoadEvonodeIdentity(pro_tx_hash, private_key)),
-                block: true,
-            },
-            status => status.into(),
-        }
-    }
-
-    fn form_name(&self) -> &'static str {
-        "Load Evonode Identity"
-    }
-
-    fn step_view(&mut self, frame: &mut Frame, area: Rect) {
-        self.input.view(frame, area)
-    }
-
-    fn step_name(&self) -> &'static str {
-        self.input.step_name()
-    }
-
-    fn step_index(&self) -> u8 {
-        self.input.step_index()
-    }
-
-    fn steps_number(&self) -> u8 {
-        self.input.steps_number()
-    }
-}
-
-struct RegisterIdentityFormController {
-    input: TextInput<DefaultTextInputParser<f64>>,
-}
-
-impl RegisterIdentityFormController {
-    fn new() -> Self {
-        RegisterIdentityFormController {
-            input: TextInput::new("Quantity (in Dash)"),
-        }
-    }
-}
-
-impl FormController for RegisterIdentityFormController {
-    fn on_event(&mut self, event: KeyEvent) -> FormStatus {
-        match self.input.on_event(event) {
-            InputStatus::Done(amount) => FormStatus::Done {
-                task: Task::Identity(IdentityTask::RegisterIdentity(
-                    (amount * 100000000.0) as u64,
-                )),
-                block: true,
-            },
-            status => status.into(),
-        }
-    }
-
-    fn form_name(&self) -> &'static str {
-        "Identity registration"
-    }
-
-    fn step_view(&mut self, frame: &mut Frame, area: Rect) {
-        self.input.view(frame, area)
-    }
-
-    fn step_name(&self) -> &'static str {
-        "Funding amount"
-    }
-
-    fn step_index(&self) -> u8 {
-        0
-    }
-
-    fn steps_number(&self) -> u8 {
-        1
-    }
-}
-
-struct TopUpIdentityFormController {
-    input: TextInput<DefaultTextInputParser<f64>>,
-}
-
-impl TopUpIdentityFormController {
-    fn new() -> Self {
-        TopUpIdentityFormController {
-            input: TextInput::new("Quantity (in Dash)"),
-        }
-    }
-}
-
-impl FormController for TopUpIdentityFormController {
-    fn on_event(&mut self, event: KeyEvent) -> FormStatus {
-        match self.input.on_event(event) {
-            InputStatus::Done(amount) => FormStatus::Done {
-                task: Task::Identity(IdentityTask::TopUpIdentity((amount * 100000000.0) as u64)),
-                block: true,
-            },
-            InputStatus::Redraw => FormStatus::Redraw,
-            InputStatus::None => FormStatus::None,
-            InputStatus::Exit => FormStatus::Exit,
-        }
-    }
-
-    fn form_name(&self) -> &'static str {
-        "Identity top up"
-    }
-
-    fn step_view(&mut self, frame: &mut Frame, area: Rect) {
-        self.input.view(frame, area)
-    }
-
-    fn step_name(&self) -> &'static str {
-        "Top up amount"
-    }
-
-    fn step_index(&self) -> u8 {
-        0
-    }
-
-    fn steps_number(&self) -> u8 {
-        1
     }
 }
 
@@ -778,122 +468,6 @@ impl FormController for WithdrawFromIdentityFormController {
 
     fn step_name(&self) -> &'static str {
         "Withdrawal amount"
-    }
-
-    fn step_index(&self) -> u8 {
-        0
-    }
-
-    fn steps_number(&self) -> u8 {
-        1
-    }
-}
-
-pub struct AddPrivateKeysFormController {
-    private_keys: Vec<String>,
-    input: ComposedInput<(
-        Field<TextInput<DefaultTextInputParser<String>>>,
-        Field<SelectInput<String>>,
-    )>,
-}
-
-impl AddPrivateKeysFormController {
-    pub fn new() -> Self {
-        Self {
-            private_keys: vec![],
-            input: ComposedInput::new((
-                Field::new("Private key", TextInput::new("Private key")),
-                Field::new(
-                    "Add another private key?",
-                    SelectInput::new(vec!["No".to_string(), "Yes".to_string()]),
-                ),
-            )),
-        }
-    }
-}
-
-impl FormController for AddPrivateKeysFormController {
-    fn on_event(&mut self, event: KeyEvent) -> FormStatus {
-        match self.input.on_event(event) {
-            InputStatus::Done((private_key, another)) => {
-                self.private_keys.push(private_key);
-
-                if another == "Yes" {
-                    self.input = ComposedInput::new((
-                        Field::new("Private key", TextInput::new("Private key in WIF format")),
-                        Field::new(
-                            "Add another private key?",
-                            SelectInput::new(vec!["No".to_string(), "Yes".to_string()]),
-                        ),
-                    ));
-                    FormStatus::Redraw
-                } else {
-                    FormStatus::Done {
-                        task: Task::Identity(IdentityTask::AddPrivateKeys(
-                            self.private_keys.clone(),
-                        )),
-                        block: false,
-                    }
-                }
-            }
-            status => status.into(),
-        }
-    }
-
-    fn form_name(&self) -> &'static str {
-        "Add identity with private key"
-    }
-
-    fn step_view(&mut self, frame: &mut Frame, area: Rect) {
-        self.input.view(frame, area)
-    }
-
-    fn step_name(&self) -> &'static str {
-        self.input.step_name()
-    }
-
-    fn step_index(&self) -> u8 {
-        self.input.step_index()
-    }
-
-    fn steps_number(&self) -> u8 {
-        self.input.steps_number()
-    }
-}
-
-struct LoadIdentityByIdFormController {
-    input: TextInput<DefaultTextInputParser<String>>,
-}
-
-impl LoadIdentityByIdFormController {
-    fn new() -> Self {
-        Self {
-            input: TextInput::new("Identity base58 ID"),
-        }
-    }
-}
-
-impl FormController for LoadIdentityByIdFormController {
-    fn on_event(&mut self, event: KeyEvent) -> FormStatus {
-        match self.input.on_event(event) {
-            InputStatus::Done(identity_id) => FormStatus::Done {
-                task: Task::Identity(IdentityTask::LoadIdentityById(identity_id)),
-                block: true,
-            },
-            status => status.into(),
-        }
-    }
-
-    fn form_name(&self) -> &'static str {
-        "Add identity with private keys"
-    }
-
-    fn step_view(&mut self, frame: &mut Frame, area: Rect) {
-        self.input.view(frame, area)
-    }
-
-    fn step_name(&self) -> &'static str {
-        "Add identity by id"
     }
 
     fn step_index(&self) -> u8 {
